Made the lazy iterator generic so that it may be reused for other implementations, specifically for other pagination calls
